requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
datasets==2.19.1
dill==0.3.6
distro==1.9.0
filelock==3.14.0
Flask==3.0.3
Flask-Caching==2.3.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.354
langchain-community==0.0.20
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langsmith==0.1.67
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
multiprocess==0.70.14
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pinecone-client==3.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.2
pydantic_core==2.18.3
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
responses==0.18.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
